Avoid null refs in BlazorIgntior when Disposed
Fixes https://github.com/aspnet/AspNetCore/issues/14257
